fix: Update promo_div class name and improve robustness
The old code would completely fail if the promo pop-up had changed class name. Now it should complete the other fixes as usual and just ignore the missing div. The others were already robust since they use an array, which would just be empty if it wasn't found.

I also updated the selector to the new name of the promo div.